Updated the environment installer to diff files
This is now setup to accept a filename property when copying over
environment specific files so that the individual environment can
be specified. I will be updating this in the next commit to be part
of the flags in env.sh so that only one execution is required with
specific flags for updating the env repo vs installing / updating
a local environment from the repo.